--- relvar-core/src/query/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod basic;
-pub mod common;
